refactor(cli): move run get render helpers to run_render.go (PRD #1009 M1)
